% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{calculate_IGscale}
\alias{calculate_IGscale}
\title{Calculate cluster contribution to IG scale paramter of full conditional for sigma2}
\usage{
calculate_IGscale(yTyk, xTyk, xTx, betak, Gk)
}
\arguments{
\item{yTyk}{squared norm of data associated with cluster k}

\item{xTyk}{crossproduct of data associated with cluster k and design matrix}

\item{xTx}{crossproduct of design matrix}

\item{betak}{current value of beta parameter for cluster k}

\item{Gk}{number of genes associated with cluster k}
}
\description{
Calculate cluster contribution to IG scale paramter of full conditional for sigma2
}

