% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster_stats}
\alias{cluster_stats}
\title{Compute summary stats for groups}
\usage{
cluster_stats(k, xTy, xTx, G, V, n, z)
}
\arguments{
\item{z}{group indicators}

\item{K}{max number of groups}
}
\description{
Compute summary stats for groups
}

