% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster_sums}
\alias{cluster_sums}
\title{Compute sums for a group}
\usage{
cluster_sums(k, yTy, xTy, G, V, z)
}
\arguments{
\item{k}{which group (zero indexed)}

\item{yTy}{numeric vector length G}

\item{xTy}{numeric matrix V by G}

\item{G}{integer}

\item{V}{integer}

\item{z}{group indicators}
}
\description{
Compute sums for a group
}

