% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{compute_weights}
\alias{compute_weights}
\title{Get component weights for all genes}
\usage{
compute_weights(yTx, xTx, beta, pi, sigma2, G, K, V)
}
\arguments{
\item{yTx}{numeric vector of length G*V}

\item{xTx}{numeric vector of length V*V}

\item{beta}{numeric vector of length K*V}

\item{pi}{numeric vector of length K}

\item{sigma2}{double}

\item{G}{int}

\item{K}{int}

\item{V}{int}

\item{n}{int}
}
\description{
Get component weights for all genes
}

