% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{construct_precision_mat}
\alias{construct_precision_mat}
\title{Construct precision matrix for full conditional distribution for beta_k}
\usage{
construct_precision_mat(xTx, Gk, sigma2, lambda2, V)
}
\arguments{
\item{xTx}{Crossproduct of design matrix}

\item{Gk}{Number of genes associated with the cluster}

\item{sigma2}{Data variance}

\item{lambda2}{Prior variance}

\item{V}{dimension of beta}
}
\description{
Construct precision matrix for full conditional distribution for beta_k
}

