% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgmm.R
\name{dpgmm}
\alias{dpgmm}
\title{Do MCMC for dpgmm model}
\usage{
dpgmm(data, design, G, V, K, N, iter)
}
\arguments{
\item{data}{A column-major $\eqn{G \cdot VN}$ gene expressionvector/matrix}

\item{design}{A column-major $\eqn{V \cdot V}$ experimental design vector/matrix}

\item{G}{number of genes}

\item{V}{number of varieties/groups}

\item{K}{maximum number of clusters}

\item{N}{samples per variety/group}
}
\description{
Do MCMC for dpgmm model
}

