% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgmm_init.R
\name{dpgmm_init}
\alias{dpgmm_init}
\title{Run MCMC for DPGMMM model, utilizing pilot chain}
\usage{
dpgmm_init(data, design, lambda2, alpha, G, V, K, N, iter, init_iter)
}
\arguments{
\item{data}{A column-major $\eqn{G \cdot VN}$ gene expressionvector/matrix}

\item{design}{A column-major $\eqn{V \cdot V}$ experimental design vector/matrix}

\item{lambda2}{prior variance for base measure}

\item{alpha}{DP concentration parameter}

\item{G}{number of genes}

\item{V}{number of varieties/groups}

\item{K}{maximum number of clusters}

\item{N}{samples per variety/group}

\item{iter}{number of iterations for main chain}

\item{init_iter}{number of iterations for pilot chain}
}
\description{
Run MCMC for DPGMMM model, utilizing pilot chain
}

