% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgmm_init_w_trim.R
\name{dpgmm_init_w_trim}
\alias{dpgmm_init_w_trim}
\title{Run MCMC for DPGMMM model, utilizing pilot chain only saving samples for certain genes}
\usage{
dpgmm_init_w_trim(data, design, lambda2, alpha, a, b, G, V, K, N, iter,
  init_iter, save_index)
}
\arguments{
\item{data}{A column-major $\eqn{G \cdot VN}$ gene expressionvector/matrix}

\item{design}{A column-major $\eqn{V \cdot V}$ experimental design vector/matrix}

\item{lambda2}{prior variance for base measure}

\item{alpha}{DP concentration parameter}

\item{G}{number of genes}

\item{V}{number of varieties/groups}

\item{K}{maximum number of clusters}

\item{N}{samples per variety/group}

\item{iter}{number of iterations for main chain}

\item{init_iter}{number of iterations for pilot chain}

\item{save_index}{zero-indexed vector indicating for which genes the save parameters will be saved}
}
\description{
Run MCMC for DPGMMM model, utilizing pilot chain only saving samples for certain genes
}

