% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpgmm_return_last.R
\name{dpgmm_return_last}
\alias{dpgmm_return_last}
\title{Do MCMC for dpgmm model returning only last iteration}
\usage{
dpgmm_return_last(data, design, lambda2, alpha, a, b, G, V, K, N, iter)
}
\arguments{
\item{data}{A column-major $\eqn{G \cdot VN}$ gene expressionvector/matrix}

\item{design}{A column-major $\eqn{V \cdot V}$ experimental design vector/matrix}

\item{lambda2}{prior variance for base measure}

\item{alpha}{DP concentration parameter}

\item{G}{number of genes}

\item{V}{number of varieties/groups}

\item{K}{maximum number of clusters}

\item{N}{samples per variety/group}
}
\description{
Do MCMC for dpgmm model returning only last iteration
}

