% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpOMP.R
\docType{package}
\name{dpOMP-package}
\alias{dpOMP}
\alias{dpOMP-package}
\title{Perform MCMC for dpgmm model}
\description{
Analyze high dimensional gene expression data using hierarchical model
that makes no parametric assumptions about the distribution of gene-specific
 parameters
}
\details{
#' @useDynLib dpOMP
}

