% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dposv.R
\name{dposv}
\alias{dposv}
\title{Solves Ax = b, replacing input b with solution and A with cholesky factor}
\usage{
dposv(n, A, B)
}
\description{
Solves Ax = b, replacing input b with solution and A with cholesky factor
}

