% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_theta.R
\name{draw_theta}
\alias{draw_theta}
\title{Draw from full conditional for beta, sigma2 conditioning on Y=y, Z=z}
\usage{
draw_theta(z, yTy, xTy, xTx, G, K, V, N)
}
\description{
Draw from full conditional for beta, sigma2 conditioning on Y=y, Z=z
}

