% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_z.R
\name{draw_z}
\alias{draw_z}
\title{Draw z for all genes given the computed weights}
\usage{
draw_z(weights, G, K)
}
\arguments{
\item{weights}{row-major vector of unnormalized weights for G genes}

\item{G}{number of genes}

\item{K}{number of categories}
}
\description{
Draw z for all genes given the computed weights
}

