% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{increment_IGscale}
\alias{increment_IGscale}
\title{Calculate cluster contribution to IG scale paramter of full conditional for sigma2}
\usage{
increment_IGscale(xTyk, xTx, betak, Gk)
}
\arguments{
\item{xTyk}{Crossproduct of data associated with cluster k and design matrix}

\item{xTx}{Crossproduct of design matrix}

\item{betak}{Current value of beta parameter for cluster k}

\item{Gk}{Number of genes associated with cluster k}
}
\description{
Calculate cluster contribution to IG scale paramter of full conditional for sigma2
}

