% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_squares.R
\name{n_squares}
\alias{n_squares}
\title{Return first n square numbers}
\usage{
n_squares(n)
}
\arguments{
\item{n}{how many d'ya want?}
}
\value{
it's pretty obvious
}
\description{
Return first n square numbers
}
\examples{
n_squares(5)

}

