% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_prime.R
\name{pi_prime}
\alias{pi_prime}
\title{Update component weight for given data}
\usage{
pi_prime(yTy, yTx, xTx, beta, pi, sigma2, V)
}
\arguments{
\item{yTx}{numeric vector of length V}

\item{xTx}{numeric vector of length V*V}

\item{beta}{numeric vector of length V}

\item{pi}{positive real number}

\item{sigma2}{positive real number}

\item{V}{int}
}
\description{
Update component weight for given data
}

