% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_form.R
\name{quad_form}
\alias{quad_form}
\title{Compute xTAx}
\usage{
quad_form(A, x, dim)
}
\arguments{
\item{A}{real-valued symmetric matrix}

\item{x}{real-valued vector with dimension compatible with A}

\item{dim}{integer giving dimensions of square matrix A}
}
\value{
numeric vector of length 1
}
\description{
Compute xTAx
}

