% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcategorical.R
\name{rcategorical}
\alias{rcategorical}
\title{Get a multinomial draw given specified weights}
\usage{
rcategorical(weights)
}
\arguments{
\item{weights}{}
}
\description{
Get a multinomial draw given specified weights
}

