% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_normaleq_symm_mat.R
\name{solve_normaleq_symm_mat}
\alias{solve_normaleq_symm_mat}
\title{Solves Ax = b, replacing input b with solution and A with cholesky factor}
\usage{
solve_normaleq_symm_mat(n, A, B)
}
\description{
Solves Ax = b, replacing input b with solution and A with cholesky factor
}

