% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{transpose}
\alias{transpose}
\title{transpose a matrix}
\usage{
transpose(m, M, N)
}
\arguments{
\item{M}{integer}

\item{N}{integer}

\item{matin}{M by N matrix or M*N vector}
}
\description{
transpose a matrix
}

