% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_stats.R
\name{cluster_sums}
\alias{cluster_sums}
\title{Compute sums for groups}
\usage{
cluster_sums(k, xTy, G, V, N, z)
}
\arguments{
\item{z}{group indicators}

\item{K}{max number of groups}
}
\description{
Compute sums for groups
}

