\name{dpOMP-package}
\alias{dpOMP-package}
\alias{dpOMP}
\docType{package}
\title{
\packageTitle{dpOMP}
}
\description{
\packageDescription{dpOMP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dpOMP}
\packageIndices{dpOMP}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{dpOMP}

Maintainer: \packageMaintainer{dpOMP}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
