% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blas_func.R
\name{quad_form}
\alias{quad_form}
\title{Computes xTAx using BLAS routines dsymv and ddot (through wrappers)}
\usage{
quad_form(x, A)
}
\description{
Computes xTAx using BLAS routines dsymv and ddot (through wrappers)
}

